// src/types/props.ts
// interface DefaultProps {
//   children?: React.ReactNode;
//   className?: string;
// }

export interface IngredientItemProps {
  ingredient: string;
}

export interface CockTailItemProps {
  strDrinkThumb: string;
  strDrink: string;
  strGlass: string;
  strAlcoholic: string;
  idDrink: string;
}